requirements.txt
pydantic>=2.0.0
PyYAML>=6.0
GitPython>=3.1.0
click>=8.0.0
rich>=13.0.0
jinja2>=3.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
requests>=2.28.0